requirements.txt
pyserial==3.5
matplotlib==3.8.0
scipy==1.11.4